requirements.txt
Django==3.1.5
django-ckeditor==6.0.0
django-filter==2.4.0
Pillow==8.1.0
xhtml2pdf==0.2.5
